+++ packages/styleguide/catalog/components/actions/actionBase.js
@@ -2,7 +2,7 @@ import { color, radius, setSpace, setType, time } from "../../../utils";
 
 const actionBase = {
   graphic: `
-    ${setSpace("phx")};
+    ${setSpace("phs")};
     ${setType("x")};
     align-content: center;
     align-items: center;